+++ responsive_data_grid/lib/src/grid_state.dart
@@ -28,6 +28,7 @@ class ResponsiveDataGridState<TItem extends Object>
   final Set<TItem> _expandedItems = <TItem>{};
   final Set<TItem> _pinnedTop = <TItem>{};
   final Set<TItem> _pinnedBottom = <TItem>{};
+  final Set<TItem> _unpinned = <TItem>{};
   GridDensity density = GridDensity.standard;
   String _searchQuery = '';
   Timer? _searchDebounce;
@@ -272,6 +273,7 @@ class ResponsiveDataGridState<TItem extends Object>
   }
 
   GridRowPin pinOf(TItem item) {
+    if (_unpinned.contains(item)) return GridRowPin.none;
     if (_pinnedTop.contains(item)) return GridRowPin.top;
     if (_pinnedBottom.contains(item)) return GridRowPin.bottom;
     return widget.rowPin?.call(item) ?? GridRowPin.none;
@@ -281,6 +283,7 @@ class ResponsiveDataGridState<TItem extends Object>
 
   void pinRow(TItem item, {GridRowPin position = GridRowPin.top}) {
     setState(() {
+      _unpinned.remove(item);
       _pinnedTop.remove(item);
       _pinnedBottom.remove(item);
       if (position == GridRowPin.top) _pinnedTop.add(item);
@@ -293,6 +296,7 @@ class ResponsiveDataGridState<TItem extends Object>
     setState(() {
       _pinnedTop.remove(item);
       _pinnedBottom.remove(item);
+      _unpinned.add(item);
     });
     _notifyState();
   }